call API from BE to update listing details
